+++ qolmat/utils/data.py
@@ -4,6 +4,8 @@ import zipfile
 from math import pi
 from typing import List
 from urllib import request
+from datetime import datetime
+from distutils.util import strtobool
 
 import numpy as np
 import pandas as pd
@@ -31,6 +33,9 @@ def get_dataframes_in_folder(path: str, extension: str) -> List[pd.DataFrame]:
         for file in files:
             if extension in file:
                 list_df.append(pd.read_csv(os.path.join(folder, file)))
+            if ".tsf" in file:
+                loaded_data = convert_tsf_to_dataframe(os.path.join(folder, file))
+                return [loaded_data]
     return list_df
 
 
@@ -71,6 +76,7 @@ def get_data(
     pd.DataFrame
         requested data
     """
+    urllink1 = "https://zenodo.org/record/"
     if name_data == "Beijing":
         urllink = "https://archive.ics.uci.edu/static/public/381/"
         zipname = "beijing+pm2+5+data"
@@ -113,6 +119,50 @@ def get_data(
         n_groups_max = min(len(sizes_stations), n_groups_max)
         stations = sizes_stations.index.get_level_values("station").unique()[-n_groups_max:]
         df = df.loc[stations]
+        return df
+    elif name_data == "Monach_weather":
+        urllink = urllink1 + "4654822/files/weather_dataset.zip?download=1"
+        zipname = "weather_dataset"
+        list_loaded_data = download_data(zipname, urllink, datapath=datapath)
+        loaded_data = list_loaded_data[0]
+        df_list: List[pd.DataFrame] = []
+        for k in range(len(loaded_data)):
+            values = list(loaded_data["series_value"][k])
+            freq = "1D"
+            time_index = pd.date_range(
+                start=pd.Timestamp("01/01/2010"), periods=len(values), freq=freq
+            )
+            df_list = df_list + [
+                pd.DataFrame(
+                    {loaded_data.series_name[k] + " " + loaded_data.series_type[k]: values},
+                    index=time_index,
+                )
+            ]
+        minimum = min([len(df) for df in df_list])
+        df = pd.concat(df_list, axis=1)
+        df = df[:minimum]
+        return df
+    elif name_data == "Monach_electricity_australia":
+        urllink = urllink1 + "4659727/files/australian_electricity_demand_dataset.zip?download=1"
+        zipname = "australian_electricity_demand_dataset"
+        list_loaded_data = download_data(zipname, urllink, datapath=datapath)
+        loaded_data = list_loaded_data[0]
+        df_list = []
+        for k in range(len(loaded_data)):
+            values = list(loaded_data["series_value"][k])
+            freq = "30min"
+            time_index = pd.date_range(
+                start=loaded_data.start_timestamp[k], periods=len(values), freq=freq
+            )
+            df_list = df_list + [
+                pd.DataFrame(
+                    {loaded_data.series_name[k] + " " + loaded_data.state[k]: values},
+                    index=time_index,
+                )
+            ]
+        minimum = min([len(df) for df in df_list])
+        df = pd.concat(df_list, axis=1)
+        df = df[:minimum]
         return df
     else:
         raise ValueError(f"Data name {name_data} is unknown!")
@@ -198,10 +248,10 @@ def add_holes(df: pd.DataFrame, ratio_masked: float, mean_size: int) -> pd.DataF
 
     generator.dict_probas_out = {column: 1 / mean_size for column in df.columns}
     generator.dict_ratios = {column: 1 / len(df.columns) for column in df.columns}
-    if generator.groups == []:
-        mask = generator.generate_mask(df)
-    else:
+    if generator.groups:
         mask = df.groupby(groups, group_keys=False).apply(generator.generate_mask)
+    else:
+        mask = generator.generate_mask(df)
 
     X_with_nans = df.copy()
     X_with_nans[mask] = np.nan
@@ -277,3 +327,128 @@ def add_datetime_features(df: pd.DataFrame) -> pd.DataFrame:
     time_cos = np.cos(2 * np.pi * time.dt.dayofyear / days_in_year)
     df["time_cos"] = np.array(time_cos)
     return df
+
+
+def convert_tsf_to_dataframe(
+    full_file_path_and_name,
+    replace_missing_vals_with="NaN",
+    value_column_name="series_value",
+):
+    col_names = []
+    col_types = []
+    all_data = {}
+    line_count = 0
+    # frequency = None
+    # forecast_horizon = None
+    # contain_missing_values = None
+    # contain_equal_length = None
+    found_data_tag = False
+    found_data_section = False
+    started_reading_data_section = False
+
+    with open(full_file_path_and_name, "r", encoding="cp1252") as file:
+        for line in file:
+            # Strip white space from start/end of line
+            line = line.strip()
+
+            if line:
+                if line.startswith("@"):  # Read meta-data
+                    if not line.startswith("@data"):
+                        line_content = line.split(" ")
+                        if line.startswith("@attribute"):
+                            if len(line_content) != 3:  # Attributes have both name and type
+                                raise Exception("Invalid meta-data specification.")
+
+                            col_names.append(line_content[1])
+                            col_types.append(line_content[2])
+                        else:
+                            if len(line_content) != 2:  # Other meta-data have only values
+                                raise Exception("Invalid meta-data specification.")
+
+                            # if line.startswith("@frequency"):
+                            #     frequency = line_content[1]
+                            # elif line.startswith("@horizon"):
+                            #     forecast_horizon = int(line_content[1])
+                            # elif line.startswith("@missing"):
+                            #     contain_missing_values = bool(strtobool(line_content[1]))
+                            # elif line.startswith("@equallength"):
+                            #     contain_equal_length = bool(strtobool(line_content[1]))
+
+                    else:
+                        if len(col_names) == 0:
+                            raise Exception("Attribute section must come before data.")
+
+                        found_data_tag = True
+                elif not line.startswith("#"):
+                    if len(col_names) == 0:
+                        raise Exception(" Attribute section must come before data.")
+                    elif not found_data_tag:
+                        raise Exception("Missing @data tag.")
+                    else:
+                        if not started_reading_data_section:
+                            started_reading_data_section = True
+                            found_data_section = True
+                            all_series = []
+
+                            for col in col_names:
+                                all_data[col] = []
+
+                        full_info = line.split(":")
+
+                        if len(full_info) != (len(col_names) + 1):
+                            raise Exception("Missing attributes/values in series.")
+
+                        series = full_info[len(full_info) - 1]
+                        series = series.split(",")
+
+                        if len(series) == 0:
+                            raise Exception(" Missing values should be indicated with ? symbol")
+
+                        numeric_series = []
+
+                        for val in series:
+                            if val == "?":
+                                numeric_series.append(replace_missing_vals_with)
+                            else:
+                                numeric_series.append(float(val))
+
+                        if numeric_series.count(replace_missing_vals_with) == len(numeric_series):
+                            raise Exception(
+                                "At least one numeric value should be there in a series."
+                            )
+
+                        all_series.append(pd.Series(numeric_series).array)
+
+                        for i in range(len(col_names)):
+                            att_val = None
+                            if col_types[i] == "numeric":
+                                att_val = int(full_info[i])
+                            elif col_types[i] == "string":
+                                att_val = str(full_info[i])
+                            elif col_types[i] == "date":
+                                att_val = datetime.strptime(full_info[i], "%Y-%m-%d %H-%M-%S")
+                            else:
+                                raise Exception("Invalid attribute type.")
+
+                            if att_val is None:
+                                raise Exception("Invalid attribute value.")
+                            else:
+                                all_data[col_names[i]].append(att_val)
+
+                line_count = line_count + 1
+
+        if line_count == 0:
+            raise Exception("Empty file.")
+        if len(col_names) == 0:
+            raise Exception("Missing attribute section.")
+        if not found_data_section:
+            raise Exception("Missing series information under data section.")
+
+        all_data[value_column_name] = all_series
+        loaded_data = pd.DataFrame(all_data)
+
+        return loaded_data
+        #     frequency,
+        #     forecast_horizon,
+        #     contain_missing_values,
+        #     contain_equal_length,