+++ qolmat/utils/exceptions.py
@@ -1,11 +1,11 @@
-from typing import Any, List
+from typing import Any, List, Tuple, Type
 
 
-class KerasExtraNotInstalled(Exception):
+class PyTorchExtraNotInstalled(Exception):
     def __init__(self):
         super().__init__(
-            """Please install keras xx.xx.xx
-        pip install qolmat[keras]"""
+            """Please install torch xx.xx.xx
+        pip install qolmat[pytorch]"""
         )
 
 
@@ -28,3 +28,13 @@ class NoMissingValue(Exception):
 class SubsetIsAString(Exception):
     def __init__(self, subset: Any):
         super().__init__(f"Provided subset `{subset}` should be None or a list!")
+
+
+class NotDimension2(Exception):
+    def __init__(self, shape: Tuple[int, ...]):
+        super().__init__(f"Provided matrix is of shape {shape}, which is not of dimension 2!")
+
+
+class NotDataFrame(Exception):
+    def __init__(self, X_type: Type[Any]):
+        super().__init__(f"Input musr be a dataframe, not a {X_type}")