+++ tests/imputations/test_imputers.py
@@ -1,10 +1,11 @@
-from typing import Any
+from typing import Any, Dict
 
 import numpy as np
 import pandas as pd
 import pytest
 from sklearn.ensemble import ExtraTreesRegressor
-from sklearn.utils.estimator_checks import parametrize_with_checks
+from sklearn.utils.estimator_checks import check_estimator, parametrize_with_checks
+from qolmat.benchmark.hyperparameters import HyperValue
 
 from qolmat.imputations import imputers
 
@@ -32,6 +33,61 @@ df_groups = pd.DataFrame(
 )
 
 
+def test_hyperparameters_get_hyperparameters() -> None:
+    imputer = imputers.ImputerKNN(n_neighbors=3)
+    hyperparams = imputer.get_hyperparams("col")
+
+    assert hyperparams == {"n_neighbors": 3, "weights": "distance"}
+
+
+hyperparams_global = {
+    "lam/col1": 4.7,
+    "lam/col2": 1.5,
+    "tol": 0.07,
+    "max_iterations": 100,
+    "norm": "L1",
+}
+
+expected1 = {
+    "lam": 4.7,
+    "tau": None,
+    "mu": None,
+    "rank": None,
+    "list_etas": (),
+    "list_periods": (),
+    "tol": 0.07,
+    "norm": "L1",
+    "max_iterations": 100,
+    "period": 1,
+}
+
+expected2 = {
+    "lam": 1.5,
+    "tau": None,
+    "mu": None,
+    "rank": None,
+    "list_etas": (),
+    "list_periods": (),
+    "tol": 0.07,
+    "norm": "L1",
+    "max_iterations": 100,
+    "period": 1,
+}
+
+
+@pytest.mark.parametrize("col, expected", [("col1", expected1), ("col2", expected2)])
+def test_hyperparameters_get_hyperparameters_modified(
+    col: str, expected: Dict[str, HyperValue]
+) -> None:
+    imputer = imputers.ImputerRPCA()
+    for key, val in hyperparams_global.items():
+        setattr(imputer, key, val)
+    imputer.imputer_params = tuple(set(imputer.imputer_params) | set(hyperparams_global.keys()))
+    hyperparams = imputer.get_hyperparams(col)
+
+    assert hyperparams == expected
+
+
 @pytest.mark.parametrize(
     "imputer",
     [
@@ -46,7 +102,7 @@ df_groups = pd.DataFrame(
 @pytest.mark.parametrize(
     "df", [pd.DataFrame({"col1": [np.nan, np.nan, np.nan], "col2": [1, 2, 3]})]
 )
-def test_Imputer_fit_transform_on_nan_column(df: pd.DataFrame, imputer: imputers.Imputer) -> None:
+def test_Imputer_fit_transform_on_nan_column(df: pd.DataFrame, imputer: imputers._Imputer) -> None:
     np.testing.assert_raises(ValueError, imputer.fit_transform, df)
 
 
@@ -58,7 +114,7 @@ def test_fit_transform_not_on_pandas(df: Any) -> None:
 
 @pytest.mark.parametrize("df", [df_groups])
 def test_fit_transform_on_grouped(df: pd.DataFrame) -> None:
-    imputer = imputers.ImputerMean(groups=["col1"])
+    imputer = imputers.ImputerMean(groups=("col1",))
     result = imputer.fit_transform(df)
     expected = pd.DataFrame(
         {
@@ -72,7 +128,8 @@ def test_fit_transform_on_grouped(df: pd.DataFrame) -> None:
 @pytest.mark.parametrize("df", [df_incomplete])
 @pytest.mark.parametrize("df_oracle", [df_complete])
 def test_ImputerOracle_fit_transform(df: pd.DataFrame, df_oracle: pd.DataFrame) -> None:
-    imputer = imputers.ImputerOracle(df_oracle)
+    imputer = imputers.ImputerOracle()
+    imputer.set_solution(df_oracle)
     result = imputer.fit_transform(df)
     expected = df_oracle
     np.testing.assert_allclose(result, expected)
@@ -151,11 +208,11 @@ def test_ImputerResiduals_fit_transform(df: pd.DataFrame) -> None:
     expected = pd.DataFrame(
         {
             "col1": [i for i in range(20)],
-            "col2": [0, 0.619048, 2, 1.435374, 2] + [i for i in range(5, 20)],
+            "col2": [0, 0.953, 2, 2.061, 2] + [i for i in range(5, 20)],
         },
         index=pd.date_range("2023-04-17", periods=20, freq="D"),
     )
-    np.testing.assert_allclose(result, expected, atol=1e-6)
+    np.testing.assert_allclose(result, expected, atol=1e-3)
 
 
 @pytest.mark.parametrize("df", [df_incomplete])
@@ -164,8 +221,8 @@ def test_ImputerKNN_fit_transform(df: pd.DataFrame) -> None:
     result = imputer.fit_transform(df)
     expected = pd.DataFrame(
         {
-            "col1": [0, 1.6666666666666667, 2, 3, 1.4285714285714286],
-            "col2": [-1, 0.3333333333333333, 0.5, 0.12499999999999994, 1.5],
+            "col1": [0, 5 / 3, 2, 3, 1.4285714285714286],
+            "col2": [-1, 1 / 3, 0.5, 1 / 8, 1.5],
         }
     )
     np.testing.assert_allclose(result, expected)
@@ -178,7 +235,6 @@ def test_ImputerMICE_fit_transform(df: pd.DataFrame) -> None:
         random_state=42,
         sample_posterior=False,
         max_iter=100,
-        missing_values=np.nan,
     )
     result = imputer.fit_transform(df)
     expected = pd.DataFrame(
@@ -196,8 +252,8 @@ def test_ImputerRegressor_fit_transform(df: pd.DataFrame) -> None:
     result = imputer.fit_transform(df)
     expected = pd.DataFrame(
         {
-            "col1": [0, 1.6666666666666667, 2, 3, 1.6666666666666667],
-            "col2": [-1, 0.3333333333333333, 0.5, 0.3333333333333333, 1.5],
+            "col1": [0, 2, 2, 3, 2],
+            "col2": [-1, 0.5, 0.5, 0.5, 1.5],
         }
     )
     np.testing.assert_allclose(result, expected)
@@ -205,7 +261,7 @@ def test_ImputerRegressor_fit_transform(df: pd.DataFrame) -> None:
 
 @pytest.mark.parametrize("df", [df_timeseries])
 def test_ImputerRPCA_fit_transform(df: pd.DataFrame) -> None:
-    imputer = imputers.ImputerRPCA(columnwise=True, max_iter=100, period=2)
+    imputer = imputers.ImputerRPCA(columnwise=False, max_iterations=100, tau=1, lam=0.3)
     result = imputer.fit_transform(df)
     expected = pd.DataFrame(
         {
@@ -213,6 +269,7 @@ def test_ImputerRPCA_fit_transform(df: pd.DataFrame) -> None:
             "col2": [0, 1, 2, 2, 2] + [i for i in range(5, 20)],
         }
     )
+    result = np.around(result)
     np.testing.assert_allclose(result, expected, atol=1e-2)
 
 
@@ -223,7 +280,7 @@ def test_ImputerEM_fit_transform(df: pd.DataFrame) -> None:
     expected = pd.DataFrame(
         {
             "col1": [i for i in range(20)],
-            "col2": [0, 1.36, 2, 4.23, 2] + [i for i in range(5, 20)],
+            "col2": [0, 21.074, 2, -4.262, 2] + [i for i in range(5, 20)],
         }
     )
     np.testing.assert_allclose(result, expected, atol=1e-2)
@@ -231,8 +288,8 @@ def test_ImputerEM_fit_transform(df: pd.DataFrame) -> None:
 
 @parametrize_with_checks(
     [
-        imputers.Imputer(),
-        imputers.ImputerOracle(df_complete),
+        imputers._Imputer(),
+        imputers.ImputerOracle(),
         imputers.ImputerMean(),
         imputers.ImputerMedian(),
         imputers.ImputerMode(),
@@ -240,14 +297,14 @@ def test_ImputerEM_fit_transform(df: pd.DataFrame) -> None:
         imputers.ImputerLOCF(),
         imputers.ImputerNOCB(),
         imputers.ImputerInterpolation(),
-        imputers.ImputerResiduals(period=7),
+        imputers.ImputerResiduals(period=2),
         imputers.KNNImputer(),
         imputers.ImputerMICE(),
         imputers.ImputerRegressor(),
-        imputers.ImputerRPCA(),
+        imputers.ImputerRPCA(tau=0, lam=0),
         imputers.ImputerEM(),
     ]
 )
-def test_sklearn_compatible_estimator(estimator: imputers.Imputer, check: Any) -> None:
+def test_sklearn_compatible_estimator(estimator: imputers._Imputer, check: Any) -> None:
     """Check compatibility with sklearn, using sklearn estimator checks API."""
     check(estimator)